backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
faiss-cpu
numpy
pydantic
tqdm
sentence-transformers
pytesseract
Pillow
opencv-python-headless
speechrecognition
pydub
python-multipart